deleted nasa_emit_modules folder from notebooks/exploratory/hr folder and moved .py files from the nasa_emit_modules to the notebooks/exploratory/hr/modules folder to make the imports and functions in the 07_hr_cwc_emit notebook work
