+++ notebooks/exploratory/hr/modules/emit_tools.py
