+++ notebooks/exploratory/hr/modules/jldp_ras_funcs.py
